% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abacus.R
\name{as.matrix.abacus}
\alias{as.matrix.abacus}
\title{Convert land cover changes data into matrix format}
\usage{
\method{as.matrix}{abacus}(data, project_id = 1, zone_id = 0, iteration = 0, scenario_id = 0)
}
\arguments{
\item{data}{abacus data}

\item{project_id}{selected project ID (default = 1)}

\item{zone_id}{selected zone ID (default = 0)}

\item{iteration}{selected iteration (default = 0 as baseline)}

\item{scenario_id}{selected scenario (default = 0 as baseline)}
}
\value{
land cover change in matrix format
}
\description{
Convert land cover changes data into matrix format
}
\examples{
as.matrix(abacus_data)
}
