% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abacus.R
\name{plot.abacus}
\alias{plot.abacus}
\title{Plot the abacus land cover change projection}
\usage{
\method{plot}{abacus}(data, scenario = NULL, project_id = 1)
}
\arguments{
\item{data}{abacus data}

\item{project_id}{selected project index}
}
\value{
plot
}
\description{
Plot the abacus land cover change projection
}
\examples{
plot(abacus_data)
}
