% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abacus.R
\name{read.abacus}
\alias{read.abacus}
\title{Read REDD Abacus file data into list object}
\usage{
read.abacus(filepath)
}
\value{
list of REDD Abacus data
}
\description{
Read REDD Abacus file data into list object
}
