% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abacus.R
\name{abacus}
\alias{abacus}
\title{Generate Abacus object}
\usage{
abacus(
  title = NULL,
  description = NULL,
  date1 = NULL,
  date2 = NULL,
  landcover,
  landcover_change,
  zone = NULL,
  carbonstock = NULL,
  other_emission_factor = NULL,
  scenario = NULL,
  n_iteration = 1,
  costbenefit = NULL
)
}
\arguments{
\item{costbenefit}{}
}
\description{
Generate Abacus object
}
